% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{dfCompile.cv.glmnet}
\alias{dfCompile.cv.glmnet}
\title{Create visualization matrix for a cv.glmnet object}
\usage{
dfCompile.cv.glmnet(x)
}
\arguments{
\item{x}{cv.glmnet object from library `glmnet`}
}
\value{
data frame, see details.
}
\description{
Create visualization matrix for a cv.glmnet object
}
\details{
data frame returned has the following columns:
\describe{
  \item{.log.lambda}{the log of the values of \code{lambda} used in the fit}
  \item{cross.validated.error}{error accumulated over across k folds of the
  cross validation for each \code{.log.lambda}}
  \item{cross.validation.upper.error}{\code{cross.validation.error} plus an
  estimate of the standard error of the \code{cross.validation.error}
  relative to the K-fold CV}
  \item{cross.validation.lower.error}{\code{cross.validation.error} minus an
  estimate of the standard error of the \code{cross.validation.error}
  relative to the K-fold CV}
  \item{number.non.zero}{number of non-zero coeffiecients at each
  \code{.log.lambda}}
  }
}
\examples{
cv_glmnet_object <- cv.glmnet(y = iris$Sepal.Length,
                              x = model.matrix(Sepal.Length~., data = iris))

dfCompile.cv.glmnet(cv_glmnet_object) \%>\% head
}
\seealso{
\code{broom}'s \code{\link[broom]{tidy}} and use with the
same object.
}
