% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gam.R
\name{dfCompile.Gam}
\alias{dfCompile.Gam}
\title{Creates an augmented data frame for \code{Gam} objects (for \pkg{ggplot2}
visuals)}
\usage{
dfCompile.Gam(x, residuals = NULL, terms = labels.Gam(x), ...)
}
\arguments{
\item{x}{\code{Gam} object from \code{gam} library}

\item{residuals}{if TRUE, partial deviance residuals are plotted along with
the fitted terms—default is FALSE. If residuals is a vector with the same
length as each fitted term in x, then these are taken to be the overall
residuals to be used for constructing the partial residuals.}

\item{terms}{subsets of the terms can be selected.}

\item{...}{(extra parameters, passed to dfCompile.Gam).}
}
\value{
augmented data.frame, see \code{details} for more information
}
\description{
Though a similar idea to \code{broom::}\code{\link[broom]{augment}} this returns
very different values.
}
\details{
\describe{
  \item{original data frame}{original data frame used to create lm or glm
  object}
  \item{.resid}{residuals between predicted verse actual \code{y} values}
 }
 For each variable in the model we have the following variables:
\describe{
  \item{.smooth._}{part of the predicted value from variable \code{_}
  for each observation}
  \item{se.smooth._.lower}{lower pointwise twice-standard-error values for
  each point relative to the part of the predicted value from variable
  \code{_}}
  \item{se.smooth._.upper}{the "upper" version of \code{se.smooth._.lower}}
}
}
\examples{
library(tidyverse)

gam.object <- gam::gam(Sepal.Length ~ gam::s(Sepal.Width) + Species,
                       data = iris)

dfCompile.Gam(gam.object) \%>\% head
}
