% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{dfCompile.lm}
\alias{dfCompile.lm}
\title{Creates an augmented data frame for lm and glm objects (for \pkg{ggplot2}
visuals)}
\usage{
dfCompile.lm(x, labels.id = factor(names(stats::residuals(x)), levels =
  names(stats::residuals(x))))
}
\arguments{
\item{x}{\code{lm} or \code{glm} object}

\item{labels.id}{labels for all observations}
}
\value{
augmented data.frame, see \code{details} for more information
}
\description{
Similar to a extended version of \code{broom::}\code{\link[broom]{augment}} for \code{lm}
and \code{glm} objects but with prepared for the diagnostic plots.
}
\details{
\describe{
  \item{original data frame}{original data frame used to create lm or glm
  object}
  \item{.index}{row number 1 to \code{nrow(original data)}}
  \item{.labels.id}{provides labels or strings with same names as
  \code{.index}}
  \item{.weights}{weights from the model for each observation}
  \item{.yhat}{predicted values in \code{y} terms
  (not probablities, logit probabilities, log transformed, etc), as such,
  for glm this is different then \code{fitted(x)}}
  \item{.resid}{residuals between \code{.yhat} and \code{y}}
  \item{.leverage}{leverage for each observation, corresponds to the diagonal
  of the "hat" matrix (\eqn{diag(X(X^TX)^{-1}X)}. }
  \item{.cooksd}{Cook's Distance, if \code{lm}, then we use the estimated
  standard deviation to calculate the value. Cook's Distance is the a
  "leave-one-out" based diagnostic for linear and generalized linear models
  discussed in Belsley, Kuh and Welsch (1980), Cook and Weisberg (1982), etc}
  \item{.weighted.resid}{residuals weighted by the \code{.weights}, (i.e.
  \eqn{\sqrt{ .weights } \cdot .resid})}
  \item{.std.resid}{the standardized residuals using weighted.residuals
  and scaled by leverage and the estimated standard deviation, (i.e.
  \eqn{\frac{\code{.weighted.resid}}{std.deviation * (1 - \code{.leverage})}}
  )}
  \item{.sqrt.abs.resid}{the square-root of the absolute value of the
  standardized residuals}
  \item{.pearson.resid}{pearson residuals, residuals divided by the square
  root the variance}
  \item{.std.pearson.resid}{standardized pearson residuals, (i.e.
  \eqn{\frac{\code{.pearson.resid}}{std.deviation * (1- \code{.leverage})}})}
  \item{.logit.leverage}{logit of the leverage
  (i.e. \eqn{log(\frac{x}{1-x})})}
  \item{.ordering.resid}{Index ordering of residuals (in absolute value)}
  \item{.ordering.std.resid}{Index ordering of standardized residuals
  (in absolute value)}
  \item{.ordering.cooks}{Index ordering of cook's distance}
  \item{.non.extreme.leverage}{logical vector if leverage != 1
  (for extreme cases)}
  }
}
\examples{
library(tidyverse)

lm.object <- lm(Sepal.Length ~., data = iris)

dfCompile.lm(lm.object) \%>\% head
}
