% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{dfCompile.matrix}
\alias{dfCompile.matrix}
\title{Creates a list of data frames for \code{tree} objects (for \pkg{ggplot2}
visuals)}
\usage{
dfCompile.matrix(x, type = "image", ...)
}
\arguments{
\item{x}{matrix object}

\item{type}{string of either "image" or "heatmap" the tells the function
whether to display the matrix directly as an image or to reorder the columns
and create a dendrogram. Note, the heatmap is a more general case of the 
image function call}

\item{...}{additional parameters (see Details section)}
}
\value{
returns a list of 1 data frame and 2 lists
\describe{
  \item{df}{contains a 'melted' data frame to visualize the matrix}
  \item{dendro_row}{a list of information about the row dendrogram in the 
  form of \code{\link[ggdendro]{dendro_data}}, namely a (1) a data frame
  called "segments" that contains containing the line segment data to
  create the dendrogram, (2) a data frame containing the label text data
  called "labels" - assoicated with the leafs, (3) empty list element
  labeled "leaf_labels", and (4) a "class" element detailing the type of 
  object it is representing (in this case always "dendrogram")}
  \item{dendro_col}{a similar list object like \code{dendro_row} but for the
  column dendrogram}.
  }
  
Note, if no dendrograms are required to be made, the entry for 
\code{dendro_row} and \code{dendro_col} is stored as NULL
}
\description{
TODO: get the function for reorder.dendrogram
}
\examples{
library(tidyverse)
data(iris)
iris_c <- iris \%>\% select(-Species) \%>\% cor

data_list <- dfCompile.matrix(iris_c, type = "heatmap")

data_list2 <- dfCompile.matrix(iris_c, type = "image")
}
