% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{dfCompile}
\alias{dfCompile}
\title{Diagnostic plot for objects (ggplot based)}
\usage{
dfCompile(x, ...)
}
\arguments{
\item{x}{object to visualize}

\item{...}{additional parameters, see \code{augment.__} for more
information}
}
\value{
data frame with ready to visualize
}
\description{
Diagnostic plot for objects (ggplot based)
}
