% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{dfCompile.tree}
\alias{dfCompile.tree}
\title{Creates a list of data frames for \code{tree} objects (for \pkg{ggplot2}
visuals)}
\usage{
dfCompile.tree(x, type = c("proportional", "uniform"))
}
\arguments{
\item{x}{}
}
\value{
returns a list of 3 data frames, see details
}
\description{
This function just a wrapper for \pkg{ggdendro}'s
\code{\link[ggdendro]{dendro_data}} (look at \code{help(dendro_data.tree)}),
with slight modifications.
}
\details{
\describe{
  \item{segments}{a data frame containing the line segment data}
  \item{labels}{a data frame containing the label text data}
  \item{leaf_labels}{a data frame containing the leaf label text data}
  }
}
\examples{
library(tidyverse)

tree.object <- tree::tree(Species ~., data = iris)

compile.df.list <- dfCompile.tree(tree.object)
compile.df.list \%>\% names

for (df.name in names(dfCompile.tree)) print(compile.df.list[["df.name"]])
}
