% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDiagnose.glmnet.R
\name{ggDiagnose.cv.glmnet}
\alias{ggDiagnose.cv.glmnet}
\title{Diagnostic plot for cv.glmnet object (ggplot based)}
\usage{
ggDiagnose.cv.glmnet(x, sign.lambda = 1, color = "red",
  dashed_color = "red", show_plot = TRUE, return = FALSE, ...)
}
\arguments{
\item{x}{cv.glmnet object from library `glmnet`}

\item{sign.lambda}{if you'd like the sign of lambda \(\\pm 1\)}

\item{color}{color of CV MSE average dot}

\item{dashed_color}{color of dashed lines for 1se and min lambda values}

\item{show_plot}{logic to display the plot}

\item{return}{logic to return list the graphic and the data frame to make
the majority of the graphic}

\item{...}{extra attributes (currently not used)}
}
\value{
depending on \code{show_plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the data frame used the make the majority of the graphic and
the graphic object itself.
}
\description{
This function creates a \code{link[ggplot2]{ggplot}} graphic version of
\code{\link[glmnet]{plot.cv.glmnet}} for \code{\link[glmnet]{cv.glmnet}}
objects.
}
