% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gam.R
\name{ggDiagnose.Gam}
\alias{ggDiagnose.Gam}
\title{Diagnostic plot for \code{Gam} object (\code{ggplot2} based)}
\usage{
ggDiagnose.Gam(x, residuals = NULL, rugplot = TRUE, se = TRUE,
  terms = labels.Gam(x), ..., jitter = TRUE, point.shape = 1,
  color = "black", se.color = "black", se.alpha = 0.2,
  se.linetype = 2, show.plot = TRUE, return = FALSE, nrow = "sqrt")
}
\arguments{
\item{x}{\code{Gam} object from \code{gam} library}

\item{residuals}{if TRUE, partial deviance residuals are plotted along with
the fitted terms—default is FALSE. If residuals is a vector with the same
length as each fitted term in x, then these are taken to be the overall
residuals to be used for constructing the partial residuals.}

\item{rugplot}{if TRUE (the default), a univariate histogram or rugplot is
displayed along the base of each plot, showing the occurrence of each ‘x’;
ties are broken by jittering (see parameter option below.}

\item{se}{if TRUE (the default), upper and lower pointwise
twice-standard-error curves are included for each plot.}

\item{terms}{subsets of the terms can be selected.}

\item{...}{(extra parameters, passed to dfCompile.Gam).}

\item{jitter}{if TRUE (the default), rug points and plotted residual points
are jittered. (Note \code{ggplot2}'s jitter is different than base plot.)}

\item{point.shape}{shape of points (the default is 1, an open circle).}

\item{color}{color for curves (or points if variable is discrete).}

\item{se.color}{color of se band.}

\item{se.alpha}{opacity of ribbon of se band if variable is continuous.}

\item{se.linetype}{linetype of se band if variable is discrete.}

\item{show.plot}{if TRUE, a display the graphics (group of graphics in this
case) is returned.}

\item{return}{if TRUE, a list of graphics and the data frame to make
the majority of graphics is returned.}

\item{nrow}{integer number of rows in the displayed graphic. (Default is
\code{"sqrt"}, which makes the \code{show.plot} output on a square grid)}
}
\value{
depending on \code{show.plot} and \code{return} it
will return the visualization of the graphics and/or a list
of both the data frame used the make the majority of the graphic and
a list of each individual graphic.
}
\description{
this function mimics the \code{\link[gam]{plot.Gam}} function from the
\pkg{gam} package, \strong{not} the \code{\link[mgcv]{plot.gam}} from
 the \pkg{mgcv} package.
}
\details{
(however thought to make 2 types of \code{gam}s is annoying.) \emph{remove
when have both.}
}
\examples{
library(gam)
gam.object <- gam::gam(Sepal.Length ~ gam::s(Sepal.Width) + Species,
                       data = iris)

par(mfrow = c(1,2))
plot(gam.object, se = TRUE, residuals = TRUE)

ggDiagnose(gam.object, residuals = TRUE) # se = TRUE by default
}
