% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDiagnose.glmnet.R
\name{ggDiagnose.glmnet}
\alias{ggDiagnose.glmnet}
\title{Diagnostic plot for glmnet object (ggplot based)}
\usage{
ggDiagnose.glmnet(x, xvar = c("norm", "lambda", "dev"), label = FALSE,
  show_plot = TRUE, return = FALSE, ...)
}
\arguments{
\item{x}{glmnet object from `glmnet` library}

\item{xvar}{string for x axis variable (see details)}

\item{label}{logic to label each beta value's line with an integer value}

\item{show_plot}{logic to display the plot}

\item{return}{logic to return list the graphic and the data frame to make
the majority of the graphic}

\item{...}{extra attributes (currently not used)}
}
\value{
depending on \code{show_plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the data frame used the make the majority of the graphic and
the graphic object itself.
}
\description{
#' This function creates a \code{link[ggplot2]{ggplot}} graphic version of
\code{\link[glmnet]{plot.glmnet}} for \code{\link[glmnet]{glmnet}}
objects.
}
