% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{ggDiagnose.matrix}
\alias{ggDiagnose.matrix}
\title{Diagnostic plot for \code{matrix} object (\code{ggplot2} based)}
\usage{
ggDiagnose.matrix(x, type = "image", dendro_height = 0.25, ...,
  show.plot = TRUE, return = FALSE)
}
\arguments{
\item{x}{matrix object}

\item{type}{string of either "image" or "heatmap" the tells the function
whether to display the matrix directly as an image or to reorder the columns
and create a dendrogram. Note, the heatmap is a more general case of the 
image function call}

\item{dendro_height}{proportional height of dendrogram compared to image
if dendrograms a displayed.}

\item{...}{used to pass additional heatmap parameters, see details in 
\code{\link{dfCompile.matrix}}.}

\item{show.plot}{logic to display the graphic}

\item{return}{logic to return list of graphics and the list of data frames to
make the majority of the graphics}
}
\value{
depending on \code{show.plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the list of data frames used the make the graphic and the
individual graphic object.
}
\description{
Creates ggplot based visual of either \code{\link[graphics]{image}} or 
\code{\link[stats]{heatmap}} for matrix objects.
}
\examples{
library(tidyverse)
data(iris)
iris_c <- iris \%>\% select(-Species) \%>\% cor \%>\% as.matrix 

ggDiagnose.matrix(iris_c, type = "heatmap")
}
