% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ggDiagnose}
\alias{ggDiagnose}
\title{Diagnostic plot for objects (ggplot based)}
\usage{
ggDiagnose(x, show.plot = TRUE, return = FALSE, ...)
}
\arguments{
\item{x}{object to visualize}

\item{show.plot}{logic to display the plot}

\item{return}{logic to return list the graphic and the data frame to make
the majority of the graphic}

\item{...}{additional parameters, see \code{ggDiagnose.__} for more
information}
}
\value{
depending on \code{show.plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the data frame used the make the majority of the graphic and
the graphic object itself.
}
\description{
Diagnostic plot for objects (ggplot based)
}
