% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{ggDiagnose.tree}
\alias{ggDiagnose.tree}
\title{Diagnostic plot for \code{tree} object (\code{ggplot2} based)}
\usage{
ggDiagnose.tree(x, type = c("proportional", "uniform"), split.labels = TRUE,
  node.labels = FALSE, text.size = 3, ..., show.plot = TRUE,
  return = FALSE)
}
\arguments{
\item{x}{an object of class \code{tree} from the \pkg{tree}.}

\item{type}{character string. If "uniform", the branches are of uniform
length (only shows depth). Otherwise they are proportional to the decrease
in impurity.}

\item{split.labels}{if TRUE (the default), non-leaf nodes are labeled with
splitting rule}

\item{node.labels}{if TRUE, leaves are labeled with predicted value
(the default is FALSE).}

\item{text.size}{integer size for \code{link[ggplot2]{geom_text}} for labels.}

\item{...}{extra attributes (currently not used)}

\item{show.plot}{logic to display the graphics (group of graphics in this
case)}

\item{return}{logic to return list of graphics and the data frame to make
the majority of graphics}
}
\value{
depending on \code{show.plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the list of data frames used the make the graphic and the
individual graphic object.
}
\description{
Utilizing work from Andie de Vries's function
\code{\link[ggdendro]{dendro_data.tree}} from the \pkg{ggdendro} package
which is used in the \code{dfCompile} to create a list of data frames.
}
\examples{
library(tree)

tree.object <- tree(Species ~., data = iris)

plot(tree.object)

ggDiagnose.tree(tree, split.labels = FALSE)
}
