% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gam.R
\name{labels.Gam}
\alias{labels.Gam}
\title{Grab term labels from \code{Gam} object}
\usage{
\method{labels}{Gam}(object, ...)
}
\arguments{
\item{object}{\code{Gam} object from package \code{gam}}

\item{...}{extra parameters (not used)}
}
\value{
term names
}
\description{
this function mimics the \code{\link[gam]{plot.Gam`}} function from the
\code{\link[gam]} package, *not* the \code{\link[mgcv]{plot.gam}} from the
\code{\link[mgcv]} package.
}
\details{
(however thought to make 2 types of \code{gam}s is annoying.) <- remove when
have both.

Inner function "lifted" from \code{gam}.
}
