% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{look.for.missing.packages}
\alias{look.for.missing.packages}
\title{Inner function to check if package names are in the namespace}
\usage{
look.for.missing.packages(package.names)
}
\arguments{
\item{package.names}{list of package names to check are in the Namespace}
}
\value{
vector of packages that are missing
}
\description{
Inner function to check if package names are in the namespace
}
