% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{look_for_missing_packages}
\alias{look_for_missing_packages}
\title{Inner function to check if package names are in the namespace}
\usage{
look_for_missing_packages(package_names)
}
\arguments{
\item{package_names}{list of package names to check are in the Namespace}
}
\value{
vector of packages that are missing
}
\description{
Inner function to check if package names are in the namespace
}
