% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{ggDiagnose.cv.glmnet}
\alias{ggDiagnose.cv.glmnet}
\title{Diagnostic plot for cv.glmnet object (ggplot based)}
\usage{
ggDiagnose.cv.glmnet(x, sign.lambda = 1, color = "red",
  dashed.color = "red", show.plot = TRUE, return = FALSE, ...)
}
\arguments{
\item{x}{\code{cv.glmnet} object from library \code{\link[glmnet]{glmnet}}}

\item{sign.lambda}{what sign of \code{lambda} you'd like \eqn{\\pm 1}}

\item{color}{color of CV MSE average dot}

\item{dashed.color}{color of dashed lines for \code{1se} and \code{min}
\code{lambda} values}

\item{show.plot}{logic to display the plot}

\item{return}{logic to return list the graphic and the data frame to make
the majority of the graphic}

\item{...}{extra attributes (currently not used)}
}
\value{
depending on \code{show.plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the data frame used the make the majority of the graphic and
the graphic object itself.
}
\description{
This function creates a \code{\link[ggplot2]{ggplot}} graphic version of
\code{\link[glmnet]{plot.cv.glmnet}} for \code{\link[glmnet]{cv.glmnet}}
objects.
}
\examples{
cv.glmnet.object <- cv.glmnet(y = iris$Sepal.Length,
                              x = model.matrix(Sepal.Length~., data = iris))

plot(cv.glmnet.object)

ggDiagnose(cv.glmnet.object)
}
\seealso{
\code{\link{dfCompile.cv.glmnet}} for data creation.
}
