% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{ggDiagnose.glmnet}
\alias{ggDiagnose.glmnet}
\title{Diagnostic plot for glmnet object (ggplot based)}
\usage{
ggDiagnose.glmnet(x, xvar = c("norm", "lambda", "dev"), label = FALSE,
  show.plot = TRUE, return = FALSE, ...)
}
\arguments{
\item{x}{\code{glmnet} object from \code{\link[glmnet]{glmnet}} library}

\item{xvar}{string for x axis variable (see details)}

\item{label}{logic to label each beta value's line with an integer value}

\item{show.plot}{logic to display the plot}

\item{return}{logic to return list the graphic and the data frame to make
the majority of the graphic}

\item{...}{extra attributes (currently not used)}
}
\value{
depending on \code{show.plot} and \code{return} it
will return the visualization of the graphic and/or a list
of both the data frame used the make the majority of the graphic and
the graphic object itself.
}
\description{
#' This function creates a \code{\link[ggplot2]{ggplot}} graphic version of
\code{\link[glmnet]{plot.glmnet}} for \code{\link[glmnet]{glmnet}}
objects.
}
\details{
For \code{xvar} you have the option of selecting
\describe{
  \item{\code{"norm"}}{The \eqn{L_1} norm of the beta values for each
  specific \code{lambda}. That is, \eqn{\sum_{i=1}^p beta_i(\lambda)}}
  \item{\code{"lambda"}}{The \code{lambda} value examined}
  \item{\code{"dev"}}{The deviance ratio, the fraction of (null) deviance
  explained by the model at each specific \code{lambda} examined. The
  deviance is defined to be \eqn{2*(loglike_sat - loglike)}, where
  \eqn{loglike_sat} is the log-likelihood for the saturated model
  (a model with a free parameter per observation). Hence
  \eqn{.dev=1-dev/nulldev}.}
}

This will be the value of the variable on the x axis.
}
\examples{
library(tidyverse)
library(glmnet)
glmnet.object <- glmnet(y = iris$Sepal.Length,
                        x = model.matrix(Sepal.Length~., data = iris))

plot(glmnet.object)

ggDiagnose.glmnet(glmnet.object)
}
\seealso{
see \code{\link{dfCompile.glmnet}} for data creation.
}
