% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gam.R
\name{labels.Gam}
\alias{labels.Gam}
\title{Grab term labels from \code{Gam} object}
\usage{
\method{labels}{Gam}(object, ...)
}
\arguments{
\item{object}{\code{Gam} object from package \code{gam}}

\item{...}{extra parameters (not used)}
}
\value{
term names
}
\description{
Inner function "lifted" from \code{gam}.
}
